+++ cypress/integration/console/devices/create-manual.spec.js
@@ -167,6 +167,35 @@ describe('End device manual create', () => {
         )
         cy.findByTestId('full-error-view').should('not.exist')
       })
+
+      it('succeeds registering a new end device without join server address', () => {
+        const device = {
+          app_eui: generateHexValue(16),
+          dev_eui: generateHexValue(16),
+          lorawan_version: 'MAC_V1_0',
+          frequency_plan_id: '863-870 MHz',
+          app_key: generateHexValue(32),
+        }
+
+        cy.findByLabelText('LoRaWAN version').selectOption(device.lorawan_version)
+        cy.findByLabelText('Frequency plan').selectOption(device.frequency_plan_id)
+        cy.findByText('Show advanced activation, LoRaWAN class and cluster settings').click()
+        cy.findByLabelText('Additional LoRaWAN class capabilities').selectOption('class-c')
+        cy.findByLabelText('Cluster settings').check()
+        cy.findByLabelText('Join Server address').clear()
+        cy.findByLabelText('DevEUI').type(device.dev_eui)
+        cy.findByLabelText('AppEUI').type(device.app_eui)
+
+        cy.findByRole('button', { name: 'Register end device' }).click()
+
+        cy.location('pathname').should(
+          'eq',
+          `${Cypress.config('consoleRootPath')}/applications/${appId}/devices/eui-${
+            device.dev_eui
+          }`,
+        )
+        cy.findByTestId('full-error-view').should('not.exist')
+      })
     })
 
     describe('LoRaWAN V1.0.2', () => {