+++ CHANGELOG.md
@@ -26,6 +26,7 @@ For details about compatibility between different releases, see the **Commitment
 
 - Emails to admins about requested OAuth clients.
 - `session` handling for joined OTAA end devices in the Console.
+- Empty Join Server address handling in end device creation form in the Console.
 
 ### Security